Complete - task 99: [fest-swing] Add methods to check properties of JTable cells http://code.google.com/p/fest/issues/detail?id=99
Created JTableCellReader (with default implementation BasicJTableCellReader) which reads the value of a cell in a JTable. BasicJTableCellReader supports JLabels, JComboBoxes and JCheckBoxes as cell renderers of a JTable.


Added methods 'fontAt(TableCell)', 'backgroundAt(TableCell)' and 'foregroundAt(TableCell)' to JTableFixture to verify the font, background and foreground of an specific cell in a JTable.