RESOLVED - issue FEST-65: JListFixture.selectItem(...) doesn'twork if list visible rectangle is very small https://kenai.com/jira/browse/FEST-65
The problem is that when selecting a list element, FEST gets the center of the cell to select. In this case, the cell is too long, and its center exceeds the width of the visible area of the JList. The solution is to detect this case, and use the x coordinate of the visible width of the JList.